Renamed isFromAndroidSdk to isInInstrumentedPackage for clarity Merge InstrumentingClassLoaderTest and abstract parent class since there is only one InstrumentingClassloader implementation. Remove two InstrumentingClassLoaderConfig subclasses by finding alternative means to configure for the tests #1 using the @Instrument annotation or mockito. Removed one now unnecessary RobolectricTestRunner subclass. Make class name translations and intercepted methods settable fields on InstrumentingClassloaderConfig. Remove two futher subclasses of InstrumentingClassloaderConfig and an overriden method in one more. Remove one more subclass of InstrumentingClassloaderConfig Remove two more test anonymous subclasses of InstrumentingClassloaderConfig overriding shouldAquire() and provide an accessor for setting classes to exclude from aquisition by the classloader. I'm not sure exactly why the tests need to do this though.
Remove last subclass of InstrumentingClassloaderConfig. When mockito is added to the whitelist two tests fail due to mockito caching mocks and attempting to reuse them between classloaders (classes are unique per classloader). I'm not sure where they are getting loaded from though, but switching to real objects rather than mocks here fixes the problem. It would be good to understand the root cause but this is likely to be more common when we run tests against multiple api levels as each one has its own classloader. The easiest way to fix is to disable the mockito mock cache which can be done by providing an implementation of org.mockito.configuration.MockitoConfiguration on the classpath and overriding enableClassCache to false (yuck!).

Migrate setter methods to a builder to create an immutable object.

Make all usages of InstrumentingClassloaderConfig go through builder.

Clean up InstrumentingClassLoaderConfig

Consolidate two fields that were doing the same thing.

Clean up some more
