+++ robolectric/src/test/java/org/robolectric/res/ResBunchTest.java
@@ -4,9 +4,8 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.TestRunners;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.fail;
-import static org.mockito.Mockito.mock;
-import static org.assertj.core.api.Assertions.*;
 import static org.robolectric.res.ResBunch.Value;
 
 @RunWith(TestRunners.WithDefaults.class)
@@ -15,9 +14,9 @@ public class ResBunchTest {
   @Test
   public void closestMatchIsPicked() {
     Values vals = new Values();
-    Value val1 = new Value("v16", mock(TypedResource.class), null);
+    Value val1 = new Value("v16", createStringTypedResource(), null);
     vals.add(val1);
-    Value val2 = new Value("v17", mock(TypedResource.class), null);
+    Value val2 = new Value("v17", createStringTypedResource(), null);
     vals.add(val2);
 
     Value v = ResBunch.pick(vals, "v18");
@@ -27,9 +26,9 @@ public class ResBunchTest {
   @Test
   public void firstValIsPickedWhenNoMatch() {
     Values vals = new Values();
-    Value val1 = new Value("en", mock(TypedResource.class), null);
+    Value val1 = new Value("en", createStringTypedResource(), null);
     vals.add(val1);
-    Value val2 = new Value("fr", mock(TypedResource.class), null);
+    Value val2 = new Value("fr", createStringTypedResource(), null);
     vals.add(val2);
 
     Value v = ResBunch.pick(vals, "v18");
@@ -39,9 +38,9 @@ public class ResBunchTest {
   @Test
   public void firstValIsPickedWhenNoQualifiersGiven() {
     Values vals = new Values();
-    Value val1 = new Value("v16", mock(TypedResource.class), null);
+    Value val1 = new Value("v16", createStringTypedResource(), null);
     vals.add(val1);
-    Value val2 = new Value("v17", mock(TypedResource.class), null);
+    Value val2 = new Value("v17", createStringTypedResource(), null);
     vals.add(val2);
 
     Value v = ResBunch.pick(vals, "");
@@ -51,9 +50,9 @@ public class ResBunchTest {
   @Test
   public void firstValIsPickedWhenNoVersionQualifiersGiven() {
     Values vals = new Values();
-    Value val1 = new Value("v16", mock(TypedResource.class), null);
+    Value val1 = new Value("v16", createStringTypedResource(), null);
     vals.add(val1);
-    Value val2 = new Value("v17", mock(TypedResource.class), null);
+    Value val2 = new Value("v17", createStringTypedResource(), null);
     vals.add(val2);
 
     Value v = ResBunch.pick(vals, "en");
@@ -63,9 +62,9 @@ public class ResBunchTest {
   @Test
   public void eliminatedValuesAreNotPickedForVersion() {
     Values vals = new Values();
-    Value val1 = new Value("en-v16", mock(TypedResource.class), null);
+    Value val1 = new Value("en-v16", createStringTypedResource(), null);
     vals.add(val1);
-    Value val2 = new Value("v17", mock(TypedResource.class), null);
+    Value val2 = new Value("v17", createStringTypedResource(), null);
     vals.add(val2);
 
     Value v = ResBunch.pick(vals, "en-v18");
@@ -75,9 +74,9 @@ public class ResBunchTest {
   @Test
   public void greaterVersionsAreNotPicked() {
     Values vals = new Values();
-    Value val1 = new Value("v11", mock(TypedResource.class), null);
+    Value val1 = new Value("v11", createStringTypedResource(), null);
     vals.add(val1);
-    Value val2 = new Value("v19", mock(TypedResource.class), null);
+    Value val2 = new Value("v19", createStringTypedResource(), null);
     vals.add(val2);
 
     Value v = ResBunch.pick(vals, "v18");
@@ -88,21 +87,21 @@ public class ResBunchTest {
   public void onlyMatchingVersionsQualifiersWillBePicked() {
     Values vals = new Values();
     
-    Value val1 = new Value("v16", mock(TypedResource.class), null);
+    Value val1 = new Value("v16", createStringTypedResource(), null);
     vals.add(val1);
-    Value val2 = new Value("sw600dp-v17", mock(TypedResource.class), null);
+    Value val2 = new Value("sw600dp-v17", createStringTypedResource(), null);
     vals.add(val2);
 
     Value v = ResBunch.pick(vals, "v18");
     assertThat(v).isEqualTo(val1);
   }
-  
+
   @Test
   public void illegalResourceQualifierThrowsException() {
     Values vals = new Values();
-    Value val1 = new Value("v11-en-v12", mock(TypedResource.class), null);
+    Value val1 = new Value("v11-en-v12", createStringTypedResource(), null);
     vals.add(val1);
-    
+
     try {
       ResBunch.pick(vals, "v18");
       fail("Expected exception to be caught");
@@ -111,7 +110,12 @@ public class ResBunchTest {
     }
   }
 
+  private static TypedResource<String> createStringTypedResource() {
+    return new TypedResource<>("title from resourceLoader1", ResType.CHAR_SEQUENCE);
+  }
+
   // Extend ResBunch.Values to make its constructor visible to this class
   private static class Values extends ResBunch.Values {
+
   }
 }
